Product repository, service, resource, GET OK

+++ backend/src/main/java/com/devsuperior/dscatalog/entities/Product.java
@@ -20,10 +20,10 @@ import javax.persistence.Table;
 @Table(name = "tb_product")
 public class Product implements Serializable{
 	private static final long serialVersionUID = 1L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	
 	@Column(columnDefinition = "TEXT")
@@ -44,7 +44,7 @@ public class Product implements Serializable{
 		
 	}
 
-	public Product(Integer id, String name, String description, Double price, String imgUrl, Instant date) {
+	public Product(Long id, String name, String description, Double price, String imgUrl, Instant date) {
 		
 		this.id = id;
 		this.name = name;
@@ -54,11 +54,11 @@ public class Product implements Serializable{
 		this.date = date;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
